refactor: 값객체 분리
