+++ src/main/java/domain/LadderBuilder.java
@@ -1,5 +1,8 @@
 package domain;
 
+import domain.values.Ladder;
+import domain.values.LinePoints;
+
 import java.util.ArrayList;
 import java.util.List;
