+++ src/main/java/domain/values/Ladder.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.values;
 
 import java.util.List;
