+++ src/main/java/view/OutputView.java
@@ -1,8 +1,5 @@
 package view;
 
-import domain.Ladder;
-
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
